Split Resend into new integrations.email namespace
