+++ frontend/src/components/icons.tsx
@@ -259,6 +259,11 @@ export const UDFIcons: Record<string, (props: CustomIconProps) => JSX.Element> =
         <Send {...rest} />
       </div>
     ),
+    "integrations.email.send_email_resend": ({ className, ...rest }) => (
+      <div className={cn("bg-lime-100", basicIconsCommon, className)}>
+        <Send {...rest} />
+      </div>
+    ),
     "core.send_email_smtp": ({ className, ...rest }) => (
       <div className={cn("bg-lime-100", basicIconsCommon, className)}>
         <Send {...rest} />
