+++ tracecat/actions/integrations/__init__.py
@@ -1,5 +1,5 @@
 from tracecat.actions.etl import extraction
 
-from . import cdr, chat, edr, enrichment, siem, sinks
+from . import cdr, chat, edr, email, enrichment, siem, sinks
 
-__all__ = ["cdr", "chat", "edr", "enrichment", "extraction", "siem", "sinks"]
+__all__ = ["cdr", "chat", "edr", "email", "enrichment", "extraction", "siem", "sinks"]
