+++ tracecat/actions/integrations/email/__init__.py
@@ -1,0 +1,5 @@
+"""Sending email."""
+
+from .resend import send_email, send_email_resend
+
+__all__ = ["send_email_resend", "send_email"]
